% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIGV.R
\name{findIGV}
\alias{findIGV}
\title{Locates the igvtools.jar or igv.jar file}
\usage{
findIGV()
}
\value{
The pathname to the igvtools.jar or igv.jar file
}
\description{
Locates the igvtools.jar or igv.jar file
}

