% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIGV.R
\name{initIGV}
\alias{findIGV}
\alias{initIGV}
\title{Initiate the Java IGV library}
\usage{
initIGV()
}
\value{
The pathname to the igvtools.jar or igv.jar file used.
}
\description{
Initiate the Java IGV library
}

